refactor: 타입 수정 및 주석 추가
