+++ src/app/(route)/customer/(route)/feedback/_hooks/useScrollToComment.ts
@@ -1,6 +1,6 @@
+// 댓글 이동 커스텀 훅
 import { useEffect } from "react";
 
-// 댓글 이동 커스텀 훅
 export function useScrollToComment(searchParams: URLSearchParams) {
   useEffect(() => {
     const commentId = searchParams.get("commentId");
