+++ src/app/(route)/customer/(route)/feedback/feedback-info/[feedbackId]/_components/StatusSaver.tsx
@@ -41,7 +41,7 @@ const statusOptions = [
 type FeedbackStatus = "PENDING" | "RECEIVED" | "COMPLETED";
 
 interface StatusSaverProps {
-  id: number;
+  id: string | string[];
   status: FeedbackStatus;
 }
